Remove schema.ts compatibility layer and use direct imports
- Migrate remaining schemas to appropriate files
- PromptSchema moved to query-schemas.ts
- Config and DatasetConfig schemas moved to base-schemas.ts
- Delete schema.ts, now fully replaced by specialized schema files
- Ensure all imports point to new schema files
- Verify all tests pass with the new structure
